requirements.txt

apache-airflow-providers-apache-spark
psycopg2-binary
Faker
